requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.10.11
aiofiles==24.1.0
pydantic==2.9.2
certifi==2024.8.30